backend_fast/auth_service/pyproject.toml
[project]
name = "auth-service"
version = "0.1.0"
description = ""
authors = [
    {name = "PolyakovVictor",email = "polyakov.victor21@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "uvicorn (>=0.34.2,<0.35.0)",
    "python-jose (>=3.4.0,<4.0.0)",
    "passlib (>=1.7.4,<2.0.0)",
    "pydantic (>=2.11.3,<3.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "bcrypt (>=4.3.0,<5.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "sqlalchemy (>=2.0.40,<3.0.0)",
    "alembic (>=1.15.2,<2.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
